# docs/backend/openapi.yml
openapi: 3.1.0
info:
  title: Practice Catalog API
  description: |
    Бэкенд мобильного каталога товаров (Добрушский фарфоровый завод).

    Соглашения:
    - Деньги — целые копейки (BYN), `integer`/`int64`.
    - Даты и время — ISO-8601, UTC (`2026-08-17T12:00:00Z`).
    - Ошибки — единый формат RFC 9457 Problem Details (схема `ProblemDetails`).
    - Пагинация — `page` (0-based) + `size` (по умолчанию 20, максимум 100); конверт `{items, page, size, total, totalPages}`.
    - Идемпотентность запросов, меняющих состояние, — заголовок `X-Request-Id`.

    Подробнее: [03-api-design.md](03-api-design.md), [02-domain-model.md](02-domain-model.md).
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальная разработка
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/orders/{id}:
    get:
      tags: [Orders]
      summary: Детали заказа
      description: Доступен только владельцу заказа.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Заказ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
              examples:
                order:
                  $ref: "#/components/examples/OrderResponseExample"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/admin/orders:
    get:
      tags: [Admin]
      summary: Все заказы
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            type: string
            enum: [NEW, CONFIRMED, DELIVERED, CANCELLED]
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Страница заказов (карточки без позиций; позиции — в деталях заказа)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrdersPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/orders/{id}:
    get:
      tags: [Admin]
      summary: Детали заказа
      description: |
        Полные данные для оператора: контакты, адрес, комментарий, позиции заказа (снимки)
        и история смены статусов.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Детали заказа
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetail"
              examples:
                detail:
                  value:
                    id: 9f2c8d1e-4b3a-4c1e-9f2c-000000000001
                    number: ORD-2026-000123
                    userId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                    status: CONFIRMED
                    itemsTotalCents: 102000
                    deliveryCents: 0
                    totalCents: 102000
                    customerName: Иванов Иван
                    customerPhone: "+375291234567"
                    deliveryCity: Добруш
                    deliveryAddress: "ул. Советская, 1"
                    comment: "Позвонить за час до доставки"
                    items:
                      - id: oi1
                        skuId: s1
                        productName: "Чайный сервиз «Славянский» на 6 персон"
                        skuName: На 6 персон
                        article: CS-1001-6
                        priceCents: 120000
                        priceWithDiscountCents: 102000
                        quantity: 1
                        totalCents: 102000
                    statusHistory:
                      - status: NEW
                        byUserId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                        at: "2026-08-17T12:00:00Z"
                      - status: CONFIRMED
                        byUserId: aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee
                        at: "2026-08-17T14:30:00Z"
                    createdAt: "2026-08-17T12:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/orders/{id}/status:
    patch:
      tags: [Admin]
      summary: Сменить статус заказа
      description: |
        Только допустимые переходы: NEW→CONFIRMED, NEW→CANCELLED,
        CONFIRMED→DELIVERED, CONFIRMED→CANCELLED. Отмена возвращает остатки.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdateRequest"
            examples:
              confirm:
                value: { status: CONFIRMED }
      responses:
        "200":
          description: Статус изменён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-токен, выданный через /api/v1/auth/login или /api/v1/auth/refresh
  examples:
...
    OrderResponseExample:
      value:
        id: 9f2c8d1e-4b3a-4c1e-9f2c-000000000001
        number: ORD-2026-000123
        status: NEW
        itemsTotalCents: 102000
        deliveryCents: 0
        totalCents: 102000
        items:
          - id: oi1
            skuId: s1
            productName: "Чайный сервиз «Славянский» на 6 персон"
            skuName: На 6 персон
            article: CS-1001-6
            priceCents: 120000
            priceWithDiscountCents: 102000
            quantity: 1
            totalCents: 102000
        createdAt: "2026-08-17T12:00:00Z"
  responses:
    ValidationError:
      description: Ошибка валидации полей запроса
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetails"
    Unauthorized:
      description: Отсутствует, просрочен или невалиден токен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: Недостаточно прав (роль)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: Конфликт состояния (409)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
    ProblemDetails:
      type: object
      description: Ошибка в формате RFC 9457
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: https://api.example.com/errors/not-found
        title:
          type: string
          example: Ресурс не найден
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: "Товар с id 123 не найден"
        instance:
          type: string
          example: /api/v1/products/123
        traceId:
          type: string
          description: Сквозной идентификатор запроса (совпадает с X-Request-Id)
          example: f7a2c9d1-4b3a-4c1e-9f2c-8d1e4b3a4c1e
    ValidationProblemDetails:
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          properties:
            errors:
              type: object
              description: "Карта ошибок валидации: поле → список сообщений"
              additionalProperties:
                type: array
                items:
                  type: string
              example:
                email: ["должен быть валидным email"]
                password: ["длина от 8 до 72 символов"]
    PageInfo:
      type: object
      description: Метаданные пагинации
      required: [page, size, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    OrderItemResponse:
      type: object
      description: Снимок товара на момент оформления
      required: [id, skuId, productName, skuName, article, priceCents, priceWithDiscountCents, quantity, totalCents]
      properties:
        id: { type: string, format: uuid }
        skuId: { type: string, format: uuid, nullable: true }
        productName: { type: string, maxLength: 300 }
        skuName: { type: string, maxLength: 200 }
        article: { type: string, maxLength: 64 }
        priceCents: { type: integer, format: int64, description: Цена на момент оформления }
        priceWithDiscountCents: { type: integer, format: int64 }
        quantity: { type: integer, minimum: 1 }
        totalCents: { type: integer, format: int64 }
    OrderResponse:
      type: object
      required: [id, number, status, itemsTotalCents, deliveryCents, totalCents, items, createdAt]
      properties:
        id: { type: string, format: uuid }
        number: { type: string, example: ORD-2026-000123 }
        status:
          type: string
          enum: [NEW, CONFIRMED, DELIVERED, CANCELLED]
        itemsTotalCents: { type: integer, format: int64 }
        deliveryCents: { type: integer, format: int64, description: В v1 всегда 0 (бесплатная доставка) }
        totalCents: { type: integer, format: int64 }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
        createdAt:
          type: string
          format: date-time
    StatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [NEW, CONFIRMED, DELIVERED, CANCELLED]
          description: Целевой статус; только допустимые переходы
...
    StatusHistoryEntry:
      type: object
      description: Запись истории смены статуса заказа
      required: [status, byUserId, at]
      properties:
        status:
          type: string
          enum: [NEW, CONFIRMED, DELIVERED, CANCELLED]
          description: Установленный статус
        byUserId:
          type: string
          format: uuid
          description: Кто выполнил переход (пользователь или оператор)
        at:
          type: string
          description: Момент перехода (ISO-8601)
          example: "2026-08-17T14:30:00Z"
    AdminOrderListItem:
      type: object
      description: Карточка заказа в списке оператора (без позиций)
      required: [id, number, userId, status, itemsTotalCents, deliveryCents, totalCents, customerName, customerPhone, itemCount, createdAt]
      properties:
        id: { type: string, format: uuid }
        number: { type: string, example: ORD-2026-000123 }
        userId: { type: string, format: uuid }
        status:
          type: string
          enum: [NEW, CONFIRMED, DELIVERED, CANCELLED]
        itemsTotalCents: { type: integer, format: int64 }
        deliveryCents: { type: integer, format: int64 }
        totalCents: { type: integer, format: int64 }
        customerName: { type: string, maxLength: 200 }
        customerPhone: { type: string, maxLength: 30 }
        itemCount:
          type: integer
          minimum: 1
          description: Число позиций в заказе (сами позиции — в деталях)
        createdAt:
          type: string
          format: date-time
    AdminOrdersPage:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AdminOrderListItem"
    AdminOrderDetail:
      type: object
      description: Полные данные заказа для оператора
      required: [id, number, userId, status, itemsTotalCents, deliveryCents, totalCents, customerName, customerPhone, deliveryCity, deliveryAddress, items, statusHistory, createdAt]
      properties:
        id: { type: string, format: uuid }
        number: { type: string, example: ORD-2026-000123 }
        userId: { type: string, format: uuid }
        status:
          type: string
          enum: [NEW, CONFIRMED, DELIVERED, CANCELLED]
        itemsTotalCents: { type: integer, format: int64 }
        deliveryCents: { type: integer, format: int64, description: В v1 всегда 0 }
        totalCents: { type: integer, format: int64 }
        customerName: { type: string, maxLength: 200 }
        customerPhone: { type: string, maxLength: 30 }
        deliveryCity: { type: string, maxLength: 100 }
        deliveryAddress: { type: string, maxLength: 300 }
        comment: { type: string, nullable: true, maxLength: 1000 }
        items:
          type: array
          description: Позиции заказа (снимки на момент оформления)
          items:
            $ref: "#/components/schemas/OrderItemResponse"
        statusHistory:
          type: array
          description: Хронология смены статусов
          items:
            $ref: "#/components/schemas/StatusHistoryEntry"
        createdAt:
          type: string
          format: date-time
...